--- tareas/ADA8/broker/src/main/java/ada8/broker/dominio/mensaje/MensajeMapper.java
@@ -1,14 +1,0 @@
-package ada8.broker.dominio.mensaje;
-
-public class MensajeMapper {
-    public static Mensaje deJsonAObjecto(String json) {
-
-        return new Mensaje();
-    }
-
-    public static String deObjetoAJson(Mensaje mensjae) {
-
-        return "En el nombre del padre, del hijo y del espiritu santo";
-    }
-
-}
