--- tareas/ADA8/broker/src/main/java/ada8/broker/dominio/mensaje/Variable.java
@@ -1,23 +1,0 @@
-package ada8.broker.dominio.mensaje;
-
-public class Variable {
-    private String variable;
-    private String valor;
-
-    public String getVariable() {
-        return variable;
-    }
-
-    public void setVariable(String variable) {
-        this.variable = variable;
-    }
-
-    public String getValor() {
-        return valor;
-    }
-
-    public void setValor(String valor) {
-        this.valor = valor;
-    }
-
-}
