--- tareas/ADA8/src/main/java/com/ada8/App.java
@@ -1,13 +1,0 @@
-package com.ada8;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
