+++ tareas/ADA8/utils/src/main/java/ada8/utilidades/Mensaje.java
@@ -1,4 +1,4 @@
-package ada8.utils.mapper;
+package ada8.utilidades;
 
 import java.util.ArrayList;
