+++ tareas/ADA8/utils/src/main/java/ada8/utilidades/MensajeMapeador.java
@@ -1,4 +1,4 @@
-package ada8.utils.mapper;
+package ada8.utilidades;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
