+++ tareas/ADA8/utils/src/main/java/ada8/utilidades/MensajeTipo.java
@@ -1,4 +1,4 @@
-package ada8.utils.mapper;
+package ada8.utilidades;
 
 public enum MensajeTipo {
     PETICION, RESPUESTA
