--- tareas/ADA8/utils/src/main/java/ada8/utils/App.java
@@ -1,13 +1,0 @@
-package ada8.utils;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!"+1 );
-    }
-}
